Moved one report endpoints to separate controller
